Last refactoring :-) / refactor due feedback
