+++ src/Module/WellKnown/HostMeta.php
@@ -18,22 +18,22 @@ class HostMeta extends BaseModule
 		$app = self::getApp();
 		$config = $app->getConfig();
 
-		header("Content-type: text/xml");
+		header('Content-type: text/xml');
 
 		if (!$config->get('system', 'site_pubkey', false)) {
 			$res = Crypto::newKeypair(1024);
 
-			$config->set('system','site_prvkey', $res['prvkey']);
-			$config->set('system','site_pubkey', $res['pubkey']);
+			$config->set('system', 'site_prvkey', $res['prvkey']);
+			$config->set('system', 'site_pubkey', $res['pubkey']);
 		}
 
 		$tpl = Renderer::getMarkupTemplate('xrd_host.tpl');
 		echo Renderer::replaceMacros($tpl, [
-				'$zhost' => $app->getHostName(),
-				'$zroot' => $app->getBaseURL(),
-				'$domain' => $app->getBaseURL(),
-				'$bigkey' => Salmon::salmonKey($config->get('system', 'site_pubkey'))]
-		);
+			'$zhost'  => $app->getHostName(),
+			'$zroot'  => $app->getBaseURL(),
+			'$domain' => $app->getBaseURL(),
+			'$bigkey' => Salmon::salmonKey($config->get('system', 'site_pubkey'))
+		]);
 
 		exit();
 	}
